+++ src/chiapos/chacha8.h
@@ -15,7 +15,7 @@ void chacha8_keysetup(struct chacha8_ctx *x, const uint8_t *k, uint32_t kbits, c
 void chacha8_get_keystream(
     const struct chacha8_ctx *x,
     uint64_t pos,
-    uint32_t n_blocks,
+    uint64_t n_blocks,
     uint8_t *c);
 
 #ifdef __cplusplus
